Bug fix
Check URL and Title when determining if a song is the same. Some services,
e.g. Spotify, advertise the same URL for every song in the album.
Also check if the same person has already been listed in another field
for the song before writing to the database, e.g. appearing in the
composer and artist fields.